Dm 2236 (#296)
* dm-2236

* dm-2236

display nulls last when order is descending.

* dm-2236

Update unit test for admin/practices page.

* dm-2236

No need for title param... Active Admin defaults to property name with correct captialization ...etc.

* dm2236

Triggering tests to be run.

* dm-2236

Remove check for unapproved practice with a valid user email... for publishing.  Going forward practices should not fall in to this state and NOT be able to be published.