dm-2800
Fixed issue in visn-show page where categories not being incremented when 'Practices Adopted in this VISN' radio button selected.  Also fixed issue with Category being incremented.. when text input text equals a category.. and either radio button ('Practices created in this VISN" OR 'Practices adopted in this VISN") is clicked.  The radio click was submitting the form.. which was causing the category to be incremented.. added a boolean flag to not when the radio is button is changed/clicked and not updating category count.